requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
Django==5.1.1
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.2
